GM2Godot refactored
Placed conversions in src/conversions
Placed GUI elements in src/gui
Versions are now changed in scr/version.py
Refactored main.py, still needs more work
