+++ src/conversion/sprites.py
