+++ src/gui/release_notes.py
@@ -1,0 +1,2 @@
+class ReleaseNotes:
+    print("WIP!")
