+++ src/gui/settings.py
@@ -1,0 +1,2 @@
+class Settings:
+    print("WIP!")
