+++ src/version.py
@@ -1,0 +1,4 @@
+VERSION = "0.0.10"
+
+def get_version():
+    return VERSION
